requirements.txt
# Core deep learning and scientific computing
torch>=2.0.0
transformers>=4.21.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
tensorboard>=2.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.13.0

# Optional: For distributed training optimization
# apex (install separately from NVIDIA)

# Development tools (optional)
# pytest>=7.0.0
# black>=22.0.0
# isort>=5.10.0
# flake8>=5.0.0